Begin work on typing imports in Webpack modules
