+++ packages/types/src/import.d.ts
@@ -1,0 +1,5 @@
+declare module "@moonlight-mod/wp/spacepack_spacepack" {
+  import { Spacepack } from "@moonlight-mod/types/coreExtensions";
+  export const spacepack: Spacepack;
+  export default spacepack;
+}
